refactor: code into separate modules for improved modularity
